project/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-admin-thumbnails==0.2.9
django-cuser==2017.3.16
django-jazzmin==3.0.1
django-tinymce==4.1.0
gunicorn==23.0.0
mysqlclient==2.2.5
packaging==24.1
pillow==11.0.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2